requirements.txt
pytest
selenium
webdriver-manager
allure-pytest
mypy
ruff